Meet exercise constraints
This separates out the input, string output, and string concatenation. The input is handled inside an if-block that only runs when run from the command-line.
The string concatenation is happens inside its own private method. The output happens in the public method, run.

This is causing the test to fail because the last thing returned from `puts` is nil so I need to fix the tests now.