Change assert_equal to assert_output
The test was failing because @greeter.run would call `puts a_string`. When you call `puts` it actually returns nil as the last thing.
Thus the comparison became "Hello, LC, nice to meet you!" == `nil` which is obviously false.
To fix this I used the assert_output method which will capture the STDOUT of the code in the block and compare that to the regex. This is more in-line with how the application works.